chore: add documentation generation
This also splits the codebase into a 'Fp' library,
which is used by the executable for golden testing.

We need to do this, since `doc-gen4` only generates docs for libraries,
not for executables. This also fits in with the broader vision
of supplying a library for end-users to verify FP semantics.